remove: withdraw 메서드를 User 패키지에서 Auth 패키지로 이동 (#214)
